refactoring (#19)
* feat: named export Typography

* chore: load typography

* update version

* chore: update dev yml

* fix zoom

* outline text

* outline and gradation

* background blink center

* wip

* add pill and progress

* suceess build
