+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "re-geo",
-  "version": "0.0.35",
+  "version": "0.0.37",
   "main": "dist/regeo.js",
   "types": "dist/index.d.ts",
   "repository": "https://github.com/sadnessOjisan/re-geo.git",
