+++ src/components/index.ts
@@ -1,0 +1,17 @@
+export * from "./alerts/Alerts";
+export * from "./background/Background";
+export * from "./badge/Badge";
+export * from "./blink/Blink";
+export * from "./button/Button";
+export * from "./button/ButtonGroup";
+export * from "./center/Center";
+export * from "./gradationText/GradationText";
+export * from "./outlineText/OutlineText";
+export * from "./pager/Pager";
+export * from "./pill/Pill";
+export * from "./pill/PillNav";
+export * from "./progressBar/ProgressBar";
+export * from "./rotator/Rotator";
+export * from "./table/Table";
+export * from "./typography/index";
+export * from "./zoom/Zoom";
