+++ src/components/typography/H1.tsx
@@ -6,19 +6,18 @@ interface Props {
   className?: string;
 }
 
-const H2 = (props: Props) => {
+const Component = (props: Props) => {
   const { children, className } = props;
-  return <StyledH2 className={className}>{children}</StyledH2>;
+  return <h1 className={className}>{children}</h1>;
 };
 
-const StyledH2 = styled.h2`
-  color: #ffcc00;
-  font-size: 36px;
+const StyledComponent = styled(Component)`
+  color: #f00;
+  font-size: 44px;
   line-height: 40px;
-  font-family: inherit;
   font-weight: bold;
   text-rendering: optimizelegibility;
   display: block;
 `;
 
-export default H2;
+export const H1 = StyledComponent;
