+++ src/components/typography/H6.tsx
@@ -6,12 +6,12 @@ interface Props {
   className?: string;
 }
 
-const H5 = (props: Props) => {
+const Component = (props: Props) => {
   const { children, className } = props;
-  return <StyledH5 className={className}>{children}</StyledH5>;
+  return <h6 className={className}>{children}</h6>;
 };
 
-const StyledH5 = styled.h5`
+const StyledComponent = styled(Component)`
   color: #00f;
   font-size: 16px;
   line-height: 40px;
@@ -21,4 +21,4 @@ const StyledH5 = styled.h5`
   display: block;
 `;
 
-export default H5;
+export const H6 = StyledComponent;
