+++ src/components/typography/index.ts
@@ -1,0 +1,7 @@
+export * from "./H1";
+export * from "./H2";
+export * from "./H3";
+export * from "./H4";
+export * from "./H5";
+export * from "./H6";
+export * from "./P";
