+++ src/index.ts
@@ -1,25 +1,2 @@
-export * from "./components/button/Button";
-export * from "./components/text/H1";
-export * from "./components/text/H2";
-export * from "./components/text/H3";
-export * from "./components/text/H4";
-export * from "./components/text/H5";
-export * from "./components/text/H6";
-export * from "./components/misc/Badge";
-export * from "./components/misc/Pager";
-export * from "./components/misc/ProgressBar";
-export * from "./components/misc/Tabs";
+export * from "./components/index";
 export * from "./img/index";
-export * from "./components/background/Background";
-export * from "./components/font/Font";
-export * from "./components/gradationText/GradationText";
-export * from "./components/marquee/Marquee";
-export * from "./components/outlineText/OutlineText";
-export * from "./components/rotator/Rotator";
-export * from "./components/zoom/Zoom";
-export * from "./components/center/Center";
-export * from "./components/blink/Blink";
-export * from "./components/table/Table";
-export * from "./components/misc/Alerts";
-export * from "./constants/color";
-export * from "./constants/font";
